pyproject.toml
[tool.black]
line-length = 110
exclude = '''
(
  /(
      \.git
    | \.eggs
    | \.venv
    | bin
    | src
    | venv
    | node_modules
    | lib
  )/
)
'''

[tool.poetry]
name = "monopoly"
version = "0.1.0"
description = ""
authors = ["Kevin Bieri <bierike@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.6"
djangorestframework = "^3.12.2"
whitenoise = {version = "^5.2.0", extras = ["brotli"]}
django-heroku = "^0.3.1"
django-extensions = "^3.1.0"
django-constance = {version = "^2.8.0", extras = ["database"]}
paho-mqtt = "^1.5.1"
django-polymorphic = "^3.0.0"
django-ordered-model = "^3.4.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"
flake8 = "^3.8.4"
pluck = "^0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
